requirements.txt
python-telegram-bot
loguru
python-decouple
